📝 moved doc definition
