+++ scrapper/docs/filters.rst
@@ -2,6 +2,3 @@ Filters
 =======
 .. automodule:: scrapper.pages.filters
    :members:
-
-.. automodule:: scrapper.pages.filters.car_filters
-   :members:
